requirements.txt
Flask
google-generativeai
youtube-transcript-api
langchain
python-dotenv
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
yt_dlp
